refactor: split session replay helpers into focused modules
